pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiosqlite>=0.21.0",
    "fastapi[standard]>=0.115.12",
    "passlib>=1.7.4",
    "pydantic-settings>=2.9.1",
    "pyjwt>=2.10.1",
    "pytest-asyncio>=0.26.0",
    "sqlmodel>=0.0.24",
]

[tool.pytest.ini_options]
pythonpath = [".", "src"]
testpaths = ["src/tests"]
filterwarnings = [
    "ignore:'crypt' is deprecated:DeprecationWarning",
    "ignore:Accessing the 'model_fields' attribute.*:DeprecationWarning",
]

